requirements.txt
Django==1.6.2
South==0.8.4
django-allauth
django-csvimport==1.0
pillow==2.3.0
django-bootstrap-form==3.1
django-bootstrap3-datetimepicker==2.2.3